feat(data): add csafe_prompt_cfr, cvl_prompt_cfr, and cvl_line_cfr data objects
Add the dataframes csafe_prompt_cfr, cvl_prompt_cfr, and cvl_line_cfr (cfr is short for cluster fill rates) to the package by saving them as data objects in the data folder.

Add the script for making the dataframes to version control by placing it in the data-raw folder. The data-raw folder is not included in the built package.

Add internal helpers for constructing these dataframes.