feat(data): drop "-cropped.png" from docnames in cvl_prompt_cfr and ".png" from cvl_line_cfc
This commit will drop "-cropped.png" from the docnames of cvl_prompt_cfr and will drop ".png" from the docnames of cvl_line_cfc. The "-cropped.png" in the docnames cvl_prompt_cfr and ".png" in the docnames of cvl_line_cfc are not used for anthing.

The docnames in csafe_prompt_cfr do not have file extensions, so this change makes the 3 data frames consistent.

Also, without the "-cropped.png" on the end of the cvl_prompt_cfr docnames, it will be easier to cross-reference each doc with its corresponding lines in cvl_line_cfc.